requirements.txt
flask
flask-cors
python-dotenv
PyPDF2
openpyxl
pandas
gunicorn
anthropic
